+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 * Fix unit tests warnings in console. (UIDATIMP-1606)
 * Wrap value for "distributionType" field mapping with quotation marks (UIDATIMP-1644)
 * Remove steps with BigTest check from pipelines on Jenkins (UIDATIMP-1648)
+* Replace `useTenantKy` with `useOkapiKy` (UIDATIMP-1649)
 
 ### Bugs fixed:
 * Remove `selected` column from associated profiles list. (UIDATIMP-1607)